requirements.txt
llama-cpp-python>=0.1.68
regex
pytest
